refactor: extract agent prompts and tools, replace IIFE with named handler
- Move chat system prompt to src/agent/prompts/chat.ts
- Move skill review prompt and builder to src/agent/prompts/skill-review.ts
- Extract shared tools to src/agent/tools.ts (queryTools, submitReviewTool)
- Parameterize createSession with systemPrompt and tools instead of hardcoding
- reviewGameSkills now uses createSession for consistency, returning JSON text
- Replace void IIFE in discord agent plugin with named handleMessage function
- Remove unused 'agent' script from package.json
