+++ agent/cli.ts
@@ -2,12 +2,18 @@ import Anthropic from '@anthropic-ai/sdk'
 import * as readline from 'node:readline'
 import { environment } from '../src/environment'
 import { createSession } from '../src/agent/create-session'
+import { chatSystemPrompt } from '../src/agent/prompts/chat'
+import { queryTools } from '../src/agent/tools'
 
 if (!environment.ANTHROPIC_API_KEY) {
   throw new Error('ANTHROPIC_API_KEY is required')
 }
 
-const session = createSession(new Anthropic({ apiKey: environment.ANTHROPIC_API_KEY }), false)
+const session = createSession(
+  new Anthropic({ apiKey: environment.ANTHROPIC_API_KEY }),
+  chatSystemPrompt(false),
+  queryTools,
+)
 
 const rl = readline.createInterface({ input: process.stdin, output: process.stdout })
