+++ src/agent/create-session.ts
@@ -1,49 +1,9 @@
 import Anthropic from '@anthropic-ai/sdk'
 import { logger } from '../logger'
-import { systemPrompt } from './system-prompt'
 import { executeQuery, type QueryTool } from './execute-query'
 
 const MAX_HISTORY = 20
 const MAX_TOOL_ITERATIONS = 8
-
-const tools: Anthropic.Tool[] = [
-  {
-    name: 'find_documents',
-    description:
-      'Run a MongoDB find() query. Use for simple lookups — fetching a player by name, getting a specific game by number, etc.',
-    input_schema: {
-      type: 'object' as const,
-      properties: {
-        collection: {
-          type: 'string',
-          description: 'Collection name (e.g. "players", "games", "logstf.logs")',
-        },
-        filter: { type: 'object', description: 'MongoDB filter object' },
-        projection: { type: 'object', description: 'Fields to include (1) or exclude (0)' },
-        sort: { type: 'object', description: 'Sort fields and direction (1 asc, -1 desc)' },
-        limit: { type: 'number', description: 'Max documents to return (hard-capped at 25)' },
-      },
-      required: ['collection'],
-    },
-  },
-  {
-    name: 'aggregate',
-    description:
-      'Run a MongoDB aggregation pipeline. Use for statistics, grouping, averages (e.g. average DPM across games, top players by game count).',
-    input_schema: {
-      type: 'object' as const,
-      properties: {
-        collection: { type: 'string', description: 'Collection name' },
-        pipeline: {
-          type: 'array',
-          description: 'Array of aggregation stage objects (e.g. [$match, $group, $sort, $limit])',
-          items: { type: 'object' },
-        },
-      },
-      required: ['collection', 'pipeline'],
-    },
-  },
-]
 
 export interface AgentSession {
   ask(question: string): Promise<{ answer: string; inputTokens: number; outputTokens: number }>
@@ -52,7 +12,9 @@ export interface AgentSession {
 
 export function createSession(
   anthropic: Anthropic,
-  isAdmin: boolean,
+  systemPrompt: string,
+  tools: Anthropic.Tool[],
+  isAdmin = false,
   initialHistory: Anthropic.MessageParam[] = [],
 ): AgentSession {
   const history: Anthropic.MessageParam[] = [...initialHistory]
@@ -73,7 +35,7 @@ export function createSession(
       const response = await anthropic.messages.create({
         model: 'claude-opus-4-7',
         max_tokens: 1024,
-        system: systemPrompt(isAdmin),
+        system: systemPrompt,
         messages,
         tools,
       })
