+++ src/discord/plugins/agent.ts
@@ -1,10 +1,13 @@
 import fp from 'fastify-plugin'
 import Anthropic from '@anthropic-ai/sdk'
+import type { Message } from 'discord.js'
 import { client } from '../client'
 import { environment } from '../../environment'
 import { logger } from '../../logger'
 import { configuration } from '../../configuration'
 import { createSession, type AgentSession } from '../../agent/create-session'
+import { chatSystemPrompt } from '../../agent/prompts/chat'
+import { queryTools } from '../../agent/tools'
 import { isOverBudget, recordUsage, getUsage } from '../../agent/daily-budget'
 import { collections } from '../../database/collections'
 
@@ -37,10 +40,8 @@ export default fp(
     const botMessageToSession = new Map<string, string>()
     const bot = client
 
-    bot.on('messageCreate', message => {
-      if (message.author.bot) return
+    async function handleMessage(message: Message): Promise<void> {
       if (!bot.user) return
-      if (!message.mentions.has(bot.user.id)) return
 
       const question = message.content
         .replace(/<@!?\d+>/g, '')
@@ -48,79 +49,88 @@ export default fp(
         .trim()
       if (!question) return
 
-      void (async () => {
-        // For threads, check the parent channel against the enabled list
-        const effectiveChannelId = message.channel.isThread()
-          ? (message.channel.parentId ?? message.channelId)
-          : message.channelId
+      // For threads, check the parent channel against the enabled list
+      const effectiveChannelId = message.channel.isThread()
+        ? (message.channel.parentId ?? message.channelId)
+        : message.channelId
 
-        const enabledChannels = await configuration.get('agent.channels')
-        if (
-          !enabledChannels.some(
-            c => c.guildId === message.guildId && c.channelId === effectiveChannelId,
-          )
-        ) {
-          return
-        }
-
-        const isAdmin =
-          (message.member?.permissions.has('Administrator') ??
-            message.member?.permissions.has('ManageGuild')) === true
-
-        const dailyTokenBudget = await configuration.get('agent.daily_token_budget')
-        if (dailyTokenBudget !== null && (await isOverBudget(dailyTokenBudget))) {
-          const { inputTokens, outputTokens } = await getUsage()
-          logger.warn(
-            { inputTokens, outputTokens, limit: dailyTokenBudget },
-            'agent daily token budget exceeded',
-          )
-          await message.reply("I've reached my daily usage limit. Try again tomorrow.")
-          return
-        }
-
-        // Determine session key:
-        //   - thread      → thread channel ID (one session per thread)
-        //   - reply chain → follow the in-memory map back to the root session key
-        //   - new mention → this message's ID starts a fresh session
-        let sessionKey: string
-        if (message.channel.isThread()) {
-          sessionKey = message.channelId
-        } else if (message.reference?.messageId) {
-          sessionKey = botMessageToSession.get(message.reference.messageId) ?? message.id
-        } else {
-          sessionKey = message.id
-        }
-
-        let session = sessions.get(sessionKey)
-        if (!session) {
-          const saved = await collections.agentSessions.findOne({ sessionKey })
-          session = createSession(anthropic, isAdmin, saved?.history ?? [])
-          sessions.set(sessionKey, session)
-        }
-
-        void message.channel.sendTyping().catch(() => {
-          /* ignore */
-        })
-
-        const { answer, inputTokens, outputTokens } = await session.ask(question)
-        await Promise.all([
-          recordUsage(inputTokens, outputTokens),
-          collections.agentSessions.updateOne(
-            { sessionKey },
-            { $set: { history: session.getHistory(), updatedAt: new Date() } },
-            { upsert: true },
-          ),
-        ])
-        logger.info(
-          { inputTokens, outputTokens, totalToday: await getUsage() },
-          'agent token usage',
+      const enabledChannels = await configuration.get('agent.channels')
+      if (
+        !enabledChannels.some(
+          c => c.guildId === message.guildId && c.channelId === effectiveChannelId,
         )
+      ) {
+        return
+      }
 
-        for (const chunk of splitMessage(answer)) {
-          const sent = await message.reply(chunk)
-          botMessageToSession.set(sent.id, sessionKey)
-        }
-      })().catch((err: unknown) => {
+      const isAdmin =
+        (message.member?.permissions.has('Administrator') ??
+          message.member?.permissions.has('ManageGuild')) === true
+
+      const dailyTokenBudget = await configuration.get('agent.daily_token_budget')
+      if (dailyTokenBudget !== null && (await isOverBudget(dailyTokenBudget))) {
+        const { inputTokens, outputTokens } = await getUsage()
+        logger.warn(
+          { inputTokens, outputTokens, limit: dailyTokenBudget },
+          'agent daily token budget exceeded',
+        )
+        await message.reply("I've reached my daily usage limit. Try again tomorrow.")
+        return
+      }
+
+      // Determine session key:
+      //   - thread      → thread channel ID (one session per thread)
+      //   - reply chain → follow the in-memory map back to the root session key
+      //   - new mention → this message's ID starts a fresh session
+      let sessionKey: string
+      if (message.channel.isThread()) {
+        sessionKey = message.channelId
+      } else if (message.reference?.messageId) {
+        sessionKey = botMessageToSession.get(message.reference.messageId) ?? message.id
+      } else {
+        sessionKey = message.id
+      }
+
+      let session = sessions.get(sessionKey)
+      if (!session) {
+        const saved = await collections.agentSessions.findOne({ sessionKey })
+        session = createSession(
+          anthropic,
+          chatSystemPrompt(isAdmin),
+          queryTools,
+          isAdmin,
+          saved?.history ?? [],
+        )
+        sessions.set(sessionKey, session)
+      }
+
+      void (message.channel as { sendTyping(): Promise<void> }).sendTyping().catch(() => {
+        /* ignore */
+      })
+
+      const { answer, inputTokens, outputTokens } = await session.ask(question)
+      await Promise.all([
+        recordUsage(inputTokens, outputTokens),
+        collections.agentSessions.updateOne(
+          { sessionKey },
+          { $set: { history: session.getHistory(), updatedAt: new Date() } },
+          { upsert: true },
+        ),
+      ])
+      logger.info({ inputTokens, outputTokens, totalToday: await getUsage() }, 'agent token usage')
+
+      for (const chunk of splitMessage(answer)) {
+        const sent = await message.reply(chunk)
+        botMessageToSession.set(sent.id, sessionKey)
+      }
+    }
+
+    bot.on('messageCreate', message => {
+      if (message.author.bot) return
+      if (!bot.user) return
+      if (!message.mentions.has(bot.user.id)) return
+
+      void handleMessage(message).catch((err: unknown) => {
         logger.error(err, 'agent error')
         message.reply('Something went wrong.').catch(() => {
           /* ignore */
